% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{formatTimestamp}
\alias{formatTimestamp}
\title{Format timestamp from the Internet Archive's Wayback Machine URLs}
\usage{
formatTimestamp(timestamp.str, tz)
}
\arguments{
\item{timestamp.str}{A character vector, with 14 digits.}

\item{tz}{The timezone, passed to `format.POSIXct`.}
}
\description{
Format timestamp from the Internet Archive's Wayback Machine URLs
}
