% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{html_text_collapse.xml_nodeset}
\alias{html_text_collapse.xml_nodeset}
\title{Internal function to correctly parse <br> tags
Suggested by \url{https://github.com/rentrop}.
See \url{https://github.com/tidyverse/rvest/issues/175#issuecomment-266434812}.}
\usage{
\method{html_text_collapse}{xml_nodeset}(x, trim = FALSE,
  collapse = "\\n")
}
\description{
Internal function to correctly parse <br> tags
Suggested by \url{https://github.com/rentrop}.
See \url{https://github.com/tidyverse/rvest/issues/175#issuecomment-266434812}.
}
