% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{parseTimestamp}
\alias{parseTimestamp}
\title{Parse timestamp from the Wayback Machine URLs}
\usage{
parseTimestamp(url, tz)
}
\arguments{
\item{url}{The Wayback Machine URL.}

\item{tz}{The timezone, passed to \code{format.POSIXct}.}
}
\description{
Parse timestamp from the Wayback Machine URLs
}
