% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve a series of pages from the Wayback Machine and parse their content.}
\usage{
retrieve(webpage = NULL, from = NULL, to = NULL, by = NULL,
  tz = NULL, filename = NULL, archive = NULL, verbose = TRUE)
}
\arguments{
\item{webpage}{The original address of the page to parse.}

\item{from}{A character indicating when to start the series, with format `YYYY-mm-dd HH:MM:SS`.}

\item{to}{A character indicating when to end the series, with format `YYYY-mm-dd HH:MM:SS`.}

\item{by}{A character indicating how to increment the sequence,
one of `hour`, `day`, `DSTday`, `week`, `month`, `quarter` or `year`.}

\item{tz}{The timezone, passed to `format.POSIXct`.}

\item{archive}{The filaname (with extension) used to save the archive,
it will overwrite without warning!.}

\item{verbose}{Logical, do you want to print information about the requests to the Wayback Machine?}

\item{archive}{An uncompleted archive list, not tested yet!}
}
\value{
`retrieve()` doesn't return anything. It saves the archive list in the working directory.
  The list details about the requests (including the sequence) in `attr` and the parsed text in `res`.
}
\description{
`retrieve()` saves a list containing the text contained in a selected number of tags.
}
